verify isSubscribed for notes

+++ src/main/java/estm/dsic/jee/controllers/NoteController.java
@@ -2,6 +2,7 @@ package estm.dsic.jee.controllers;
 
 import java.util.List;
 import estm.dsic.jee.models.Note;
+import estm.dsic.jee.security.JwtUtil;
 import estm.dsic.jee.services.NoteService;
 import jakarta.inject.Inject;
 import jakarta.ws.rs.*;
@@ -22,9 +23,18 @@ public class NoteController {
     }
 
     @GET
-    @Path("/{id}")
+    @Path("/{id}") // this id is for the Owner id means the notes of a spicifique user
     @Produces(MediaType.APPLICATION_JSON)
-    public Response getNoteById(@PathParam("id") int id) {
+    public Response getNoteById(@PathParam("id") int id, @CookieParam("g_note_jwt") String jwtToken) {
+
+        // Check if the user is subscribed
+        Response subscriptionResponse = JwtUtil.verifyisSubscribed(jwtToken);
+        if (subscriptionResponse.getStatus() != Response.Status.OK.getStatusCode()) {
+            // If the user is not an admin or there was an issue with authentication, return
+            // the response
+            return subscriptionResponse;
+        }
+
         List<Note> notes = noteService.getNotesByOwnerId(id);
         if (notes != null) {
             return Response.ok(notes).build();
@@ -35,7 +45,16 @@ public class NoteController {
 
     @POST
     @Consumes(MediaType.APPLICATION_JSON)
-    public Response createNote(Note note) {
+    public Response createNote(Note note, @CookieParam("g_note_jwt") String jwtToken) {
+
+        // Check if the user is subscribed
+        Response subscriptionResponse = JwtUtil.verifyisSubscribed(jwtToken);
+        if (subscriptionResponse.getStatus() != Response.Status.OK.getStatusCode()) {
+            // If the user is not an admin or there was an issue with authentication, return
+            // the response
+            return subscriptionResponse;
+        }
+
         try {
             System.out.println("Note to add:\n" + note);
 
@@ -59,7 +78,16 @@ public class NoteController {
 
     @PUT
     @Consumes(MediaType.APPLICATION_JSON)
-    public Response updateNote(Note note) {
+    public Response updateNote(Note note, @CookieParam("g_note_jwt") String jwtToken) {
+
+        // Check if the user is subscribed
+        Response subscriptionResponse = JwtUtil.verifyisSubscribed(jwtToken);
+        if (subscriptionResponse.getStatus() != Response.Status.OK.getStatusCode()) {
+            // If the user is not an admin or there was an issue with authentication, return
+            // the response
+            return subscriptionResponse;
+        }
+
         try {
             // Log the note being updated
             // System.out.println("Note to update:\n" + note);
@@ -85,7 +113,16 @@ public class NoteController {
 
     @DELETE
     @Path("/{id}")
-    public Response deleteNoteById(@PathParam("id") int id) {
+    public Response deleteNoteById(@PathParam("id") int id, @CookieParam("g_note_jwt") String jwtToken) {
+        System.out.println("\n\n\ndelete note api is called ....");
+        // Check if the user is subscribed
+        Response subscriptionResponse = JwtUtil.verifyisSubscribed(jwtToken);
+        if (subscriptionResponse.getStatus() != Response.Status.OK.getStatusCode()) {
+            // If the user is not an admin or there was an issue with authentication, return
+            // the response
+            return subscriptionResponse;
+        }
+
         try {
             boolean success = noteService.delete(id);
             if (success) {
@@ -103,8 +140,18 @@ public class NoteController {
     @GET
     @Path("/search")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response searchNotes(@QueryParam("keyword") String keyword, @QueryParam("userId") int userId) {
-        System.out.println("the key word is : "+keyword);
+    public Response searchNotes(@QueryParam("keyword") String keyword, @QueryParam("userId") int userId,
+            @CookieParam("g_note_jwt") String jwtToken) {
+        System.out.println("\n\nthe key word is : " + keyword);
+
+        // Check if the user is subscribed
+        Response subscriptionResponse = JwtUtil.verifyisSubscribed(jwtToken);
+        if (subscriptionResponse.getStatus() != Response.Status.OK.getStatusCode()) {
+            // If the user is not an admin or there was an issue with authentication, return
+            // the response
+            return subscriptionResponse;
+        }
+
         try {
             if (keyword == null || keyword.isEmpty()) {
                 return Response.status(Response.Status.BAD_REQUEST)
@@ -112,13 +159,13 @@ public class NoteController {
                         .build();
             }
 
-            List<Note> notes = noteService.searchNotesByKeyword(userId,keyword);
-            System.out.println("notes found"+notes);
+            List<Note> notes = noteService.searchNotesByKeyword(userId, keyword);
+            System.out.println("notes found" + notes);
             if (!notes.isEmpty()) {
                 return Response.ok(notes).build();
             } else {
                 return Response.status(Response.Status.NOT_FOUND)
-                        .entity("No notes found for the given keyword : '"+keyword+"'")
+                        .entity("No notes found for the given keyword : '" + keyword + "'")
                         .build();
             }
         } catch (Exception e) {
